Newsletter functionality in main and about activities
Adds a newsletter sign-up option in main activity at the top
of the page, which has error and success messaging. It can
be dismissed (with dismiss messaging).

Adds newsletter sign-up to about activity using same
error and success messaging, but without dismiss option.

Only shows newsletter sections if the user hasn't signed
up yet, or in the case of the main activity, hasn't
dismissed it yet.

Fixes #125