Adds search functionality for issue reason (#192)
Searches issue reason as well as category and title.
Uses regex to ensure that word starts are matched, which
keeps "ice" from matching "averice", for example. Adds
tests for this behavior. Copies FakeJSONData to support
tests as sharing between test/ and androidtest/ seems
nontrivial, but should be done in a follow-up.