Rep call activity polish & updates (#205)
* Shows previous call stats instead of just green check in RepCallActivity

Replaces clickable green checkmark (which wasn't discoverable) with a count of previous calls and a red "details" button which shows a dialog that lists previous call results. Updates previous call results from a list of types of outcomes to a list of all previous outcomes.

* Adds contact details to rep call activity, fixes #204

Adds contact political party (if present), district (if present) and state to the rep's description in RepCallActivity. Also uses `reason` instead of `area` in IssueActivity; `area` is more of an enum and `reason` is human-readable.

* Show call outcome count instead of listing individual call outcomes