Fix crash in Settings when reminders are enabled but notifications are not enabled
registerForActivityResult must be called from onCreate or we get a crash. Fixes the crash and tests that the notifications dialog is shown. Leaves as TODO to tell the user they must go into settings to turn notifications on after notifications are disabled the first time. This could do with some polish but fixes the crash in #211